requirements.txt
flask 
beautifulsoup4 
requests 
selenium 
webdriver-manager 
pandas 
geopy 
schedule 
gunicorn
psycopg2-binary
python-dotenv
